Added local tests for get event request
